requirements.txt
langchain
langchain-community
langchain-core

langchain-groq
langchain-google-genai

faiss-cpu
fastapi
uvicorn
python-dotenv
python-multipart
PyMuPDF
structlog
docx2txt
ipykernel
streamlit
pytest
pypdf
pyyaml
-e .
